pyproject.toml
 [project]
 name = "cv-intelligence-layer"
 version = "0.1.0"
 description = "CV Intelligence Layer — FastAPI middleware for CV processing on top of Semantic Search"
 requires-python = ">=3.11"
 dependencies = [
   "fastapi>=0.115",
   "uvicorn[standard]>=0.30",
   "pydantic>=2.7",
   "pydantic-settings>=2.3",
   "sqlalchemy>=2.0",
   "asyncpg>=0.29",
   "alembic>=1.13",
   "httpx>=0.27",
   "tenacity>=8.3",
   "python-multipart>=0.0.9",
   "python-json-logger>=2.0.7",
   "celery>=5.4",
   "redis>=5.0",
   "google-genai>=1.0.0",
   "PyMuPDF>=1.24",
   "python-docx>=1.1.2",
   "easyocr>=1.7.1",
   "numpy>=1.26.0",
 ]
 
 [project.optional-dependencies]
 dev = [
   "pytest>=8.2",
   "pytest-asyncio>=0.23.7",
   "respx>=0.21.1",
   "ruff>=0.5.5",
   "mypy>=1.10",
   "types-redis>=4.6.0.20240425",
 ]
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 testpaths = ["tests"]
 
 [tool.ruff]
 line-length = 100
 target-version = "py311"
 
 [tool.mypy]
 python_version = "3.11"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 no_implicit_optional = true

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

 
